WAITP-1219 Update dockerfile
